requirements.txt
django==5.0.1
djangorestframework==3.14.0
psycopg==3.1.12
pdf2image==1.16.3
pytesseract==0.3.10
pdfplumber==0.10.2
opencv-python==4.8.1.78
pydantic==2.10.6
python-dotenv==0.21.0